requirements.txt
# Biashara Bot — Python dependencies
# Install with: pip install -r requirements.txt

# MCP (Model Context Protocol) SDK — powers the local data server
# Includes the FastMCP high-level API and CLI utilities
mcp[cli]>=1.0.0

# Pydantic — used for typed tool parameter descriptions
pydantic>=2.0.0

# Flask — web framework for the chat web app
flask>=3.0.0

# OpenAI SDK — used by the web app to call GitHub Models API
openai>=1.0.0

# python-dotenv — loads .env files for configuration
python-dotenv>=1.0.0
